Completed implementation of Power Profiler
