+++ neurio/devices/monitors/power_monitor.py
@@ -9,11 +9,171 @@ Description: Power Monitoring classes
 """
 
 import time
-from ppk2_api.ppk2_api import PPK2_API
 import threading
 import json
 import os
 import numpy as np
+
+import time
+import serial
+import struct
+import logging
+from serial.threaded import ReaderThread, Protocol
+
+class PPK2Command:
+    GET_METADATA = 0x19
+    AVERAGE_START = 0x06
+    AVERAGE_STOP = 0x07
+    RESET = 0x20
+    SET_POWER_MODE = 0x11
+    SET_USER_GAINS = 0x25
+    DEVICE_RUNNING_SET = 0x0C
+    TRIGGER_SET = 0x01
+    NO_OP = 0x00
+    REGULATOR_SET = 0x0d
+
+
+class PPK2Protocol(Protocol):
+    def __init__(self):
+        super().__init__()
+        self.transport = None
+        self.buffer = b""
+        self.sample_size = 4
+        self.ppk2 = None  # Back reference
+
+    def connection_made(self, transport):
+        self.transport = transport
+        logging.info("Serial connection established.")
+
+    def data_received(self, data):
+        self.buffer += data
+        while len(self.buffer) >= self.sample_size:
+            sample = int.from_bytes(self.buffer[:4], 'little')
+            self.buffer = self.buffer[4:]
+            if self.ppk2:
+                self.ppk2.handle_sample(sample)
+
+    def connection_lost(self, exc):
+        logging.warning("Serial connection lost.")
+
+
+class PPK2:
+    def __init__(self, port, source_voltage=5000, baudrate=115200):
+        self.port = port
+        self.baudrate = baudrate
+        self.serial_port = serial.Serial(port, baudrate=baudrate, timeout=1)
+        self.reader_thread = None
+        self.protocol = None
+
+        self.expected_counter = None
+        self.remainder = b""
+        self.recording = []
+        self.modifiers = {
+            "r": [1031.64, 101.65, 10.15, 0.94, 0.043],
+            "gs": [1, 1, 1, 1, 1],
+            "gi": [1, 1, 1, 1, 1],
+            "o": [0, 0, 0, 0, 0],
+            "s": [0, 0, 0, 0, 0],
+            "i": [0, 0, 0, 0, 0],
+            "ug": [1, 1, 1, 1, 1],
+        }
+        self.adc_mult = 1.8 / 163840
+        self.vdd_low = 800
+        self.vdd_high = 5000
+
+        self.set_source_voltage(source_voltage)
+
+        if self.reader_thread is not None:
+            logging.warning("Reader thread already running.")
+            return
+        self.reader_thread = ReaderThread(self.serial_port, PPK2Protocol)
+        self.reader_thread.start()
+        self.protocol = self.reader_thread.connect()[1]
+        self.protocol.ppk2 = self
+
+    def _convert_source_voltage(self, mV):
+        """Convert input voltage to device command"""
+        # minimal possible mV is 800
+        if mV < self.vdd_low:
+            mV = self.vdd_low
+
+        # maximal possible mV is 5000
+        if mV > self.vdd_high:
+            mV = self.vdd_high
+
+        offset = 32
+        # get difference to baseline (the baseline is 800mV but the initial offset is 32)
+        diff_to_baseline = mV - self.vdd_low + offset
+        base_b_1 = 3
+        base_b_2 = 0  # is actually 32 - compensated with above offset
+
+        # get the number of times we have to increase the first byte of the command
+        ratio = int(diff_to_baseline / 256)
+        remainder = diff_to_baseline % 256  # get the remainder for byte 2
+
+        set_b_1 = base_b_1 + ratio
+        set_b_2 = base_b_2 + remainder
+
+        return set_b_1, set_b_2
+
+
+    def set_source_voltage(self, mV):
+        """Inits device - based on observation only REGULATOR_SET is the command.
+        The other two values correspond to the voltage level.
+
+        800mV is the lowest setting - [3,32] - the values then increase linearly
+        """
+        b_1, b_2 = self._convert_source_voltage(mV)
+        self.send_command(PPK2Command.REGULATOR_SET, b_1, b_2)
+        self.current_vdd = mV
+
+    def start(self):
+        self.recording = []
+        self.send_command(PPK2Command.AVERAGE_START)
+
+    def stop(self):
+        self.send_command(PPK2Command.AVERAGE_STOP)
+
+    def close(self):
+        if self.reader_thread:
+            self.reader_thread.close()
+            self.reader_thread = None
+            self.protocol = None
+        if self.reader_thread:
+            self.reader_thread.close()
+
+    def __del__(self):
+        self.disable_DUT()
+        self.close()
+
+    def enable_DUT(self):
+        self.send_command(PPK2Command.DEVICE_RUNNING_SET, PPK2Command.TRIGGER_SET)
+
+    def disable_DUT(self):
+        self.send_command(PPK2Command.DEVICE_RUNNING_SET, PPK2Command.NO_OP)
+
+    def send_command(self, *args):
+        self.serial_port.write(bytes(args))
+
+    def handle_sample(self, adc_value: int):
+        range_ = (adc_value >> 14) & 0x7
+        counter = (adc_value >> 18) & 0x3F
+        adc_result = (adc_value & 0x3FFF) * 4
+        bits = (adc_value >> 24) & 0xFF
+
+        res = (adc_result - self.modifiers["o"][range_]) * (self.adc_mult / self.modifiers["r"][range_])
+        adc = self.modifiers["ug"][range_] * (
+            res * (self.modifiers["gs"][range_] * res + self.modifiers["gi"][range_]) +
+            (self.modifiers["s"][range_] * (self.current_vdd / 1000) + self.modifiers["i"][range_])
+        )
+
+        self.recording.append({"value": 0.5*adc * 1e6, "bits": bits, "time": time.time()})
+
+    def get_metadata(self):
+        self.send_command(PPK2Command.GET_METADATA)
+        data = self.serial_port.read(512).decode(errors='ignore')
+        print("Metadata:", data)
+        return data
 
 
 class PowerProfilerKitII:
@@ -22,24 +182,21 @@ class PowerProfilerKitII:
     """
 
     def __init__(self, port: any = "serial", name: str = "ppk2", mode="source", source_voltage: int = 5000,
-                 acquisition_interval: float = 0.01, verbose: int = 0, log_dir: str = None, limit_uA=None, **kwargs):
-        self.port = port
+                 verbose: int = 0, log_dir: str = None, limit_uA=None, **kwargs):
+
+        self.kit = PPK2(port=port, source_voltage=source_voltage, baudrate=115200)
         self.name = name
         self.recordings = []
-        self.current_phase = None
-        self.mode = mode
+        if mode not in ["source"]:
+            raise Exception(f'Unsupported mode {mode}')
         self.verbose = verbose
-        self.acquisition_interval = acquisition_interval
         if log_dir is None:
             log_dir = os.path.join("logs-{}".format(time.strftime("%Y%m%d-%H%M%S")))
         self.log_dir = log_dir
         os.makedirs(self.log_dir, exist_ok=True)
         self.limit_uA = limit_uA
-
-        # init PPK2
-        self.source_voltage = source_voltage
-        self.kit = None
-        self.__connect__()
+        self.phase = None
+        self.running = False
 
     def set_source_voltage(self, int):
         """
@@ -47,18 +204,8 @@ class PowerProfilerKitII:
         :param int: voltage in mV
         """
         self.kit.set_source_voltage(int)
-
-    def set_mode(self, mode):
-        """
-        Set the mode of the Power Profiler Kit II.
-        """
-        if mode not in ["ampere", "source"]:
-            raise Exception(f'Unknown mode {mode}')
-        if mode == "ampere":
-            self.kit.use_ampere_meter()
-        elif mode == "source":
-            self.kit.use_source_meter()
-        self.mode = mode
+        if self.verbose:
+            print(f"Source voltage set to {int} mV")
 
     def toggle_power(self, state):
         """
@@ -66,77 +213,11 @@ class PowerProfilerKitII:
         :param state: "ON" or "OFF"
         """
         if state not in ["ON", "OFF"]:
-            raise Exception(f'Unknown state {state}')
-        self.kit.toggle_DUT_power(state)
-
-    def __connect__(self):
-        ppk2s_connected = PPK2_API.list_devices()
-        print("PPK2s connected: ", ppk2s_connected)
-        if (len(ppk2s_connected) == 1):
-            ppk2_port = ppk2s_connected[0]
-            if self.verbose > 0: print(f'Found PPK2 at {ppk2_port}')
-        elif len(ppk2s_connected) > 1:
-            if self.verbose > 0: print("More than one PPK2 connected")
-            ppk2_port = self.port
-            for ppk2 in ppk2s_connected:
-                if self.port == ppk2:
-                    ppk2_port = ppk2
-                    if self.verbose > 0: print(f'Using PPK2 at {ppk2}')
-        else:
-            raise Exception("No PPK2 connected")
-
-        if self.kit:
-            del self.kit
-            time.sleep(3.0)
-        self.kit = PPK2_API(ppk2_port)
-        time.sleep(0.5)
-        self.kit.get_modifiers()
-
-        self.set_source_voltage(self.source_voltage)
-        if self.mode == "ampere":
-            self.kit.use_ampere_meter()
-        elif self.mode == "source":
-            self.kit.use_source_meter()
-        else:
-            raise Exception(f'Unknown mode {self.mode}')
-        self.kit.toggle_DUT_power("ON")  # enable DUT power
-        self.running = False
-        self.buffer = []
-        self.phase = None
-        self.failed = False
-        time.sleep(1.0)
-
-    def __record__(self):
-        """
-        Concurrently record the power consumption of the device.
-        """
-        while self.running:
-            if not self.failed:
-                read_data = self.kit.get_data()
-                if read_data != b'':
-                    samples, raw_digital = self.kit.get_samples(read_data)
-                    self.recording["samples"] += samples  # concatenate new samples
-                    self.recording["digital"] += raw_digital  # concatenate new digital data
-                    if self.limit_uA is not None:
-                        max_val = np.asarray(samples).max()
-                        if max_val > self.limit_uA:
-                            print(
-                                f"============= FAILED - Power ({max_val}) uA exceeds {self.limit_uA} uA=============")
-                            self.failed = True
-                            self.running = False  # stop recording
-                    # Raw digital contains the raw digital data from the PPK2
-                    # The number of raw samples is equal to the number of samples in the samples list
-                    # We have to process the raw digital data to get the actual digital data
-
-                    # TODO handle channels
-                    """
-                        digital_channels = self.kit.digital_channels(raw_digital)
-                        for ch in digital_channels:
-                            # Print last 10 values of each channel
-                            print(ch[-10:])
-                        print()
-                    """
-                time.sleep(self.acquisition_interval)
+            raise Exception(f'Unsupported state {state}')
+        if state == "ON":
+            self.kit.enable_DUT()
+        elif state == "OFF":
+            self.kit.disable_DUT()
 
     def start(self, phase: str):
         """
@@ -147,32 +228,23 @@ class PowerProfilerKitII:
         if self.running:
             raise Exception("Already recording. Stop first")
         else:
-            if self.failed:
-                print("Last measurement failed. Resetting the device and the PPK2")
-                self.__connect__()
-
             self.phase = phase
-            self.kit.start_measuring()  # start measuring
-            time.sleep(0.1)
-            self.recording = {"phase": self.phase, "samples": [], "digital": [], "start_timestamp": time.time()}
-            self.running = True
-            self.thread = threading.Thread(target=self.__record__, daemon=True)
-            self.thread.start()
+            self.recording = {"phase": phase, "start_timestamp": time.time(), "samples": []}
+            self.kit.start()
 
     def stop(self):
         """
         Stop recording the power consumption of the device.
         """
-        self.running = False
+        self.kit.stop()
         self.recording["stop_timestamp"] = time.time()
-        self.kit.stop_measuring()  # stop measuring
-        time.sleep(1.0)
-        if self.thread is not None:
-            self.thread.join()
-            self.recordings.append(self.recording)
-            with open(os.path.join(self.log_dir, f"{self.name}_recording.json"), "w") as f:
-                json.dump(self.recordings, f)
-            self.thread = None
-            self.recording = None
-        else:
-            raise Exception("Not recording. Start first")
+        self.running = False
+        # get samples from kit
+        times = [s['time'] - self.kit.recording[0]['time'] for s in self.kit.recording]
+        values = [s['value'] for s in self.kit.recording]
+
+        self.recording["samples"] = [{"time": t, "value": v} for t, v in zip(times, values)]
+        self.recordings.append(self.recording)
+
+        with open(os.path.join(self.log_dir, f"{self.name}_recording.json"), "w") as f:
+            json.dump(self.recordings, f)
